# [SPARK-28128][PYTHON][SQL] Pandas Grouped UDFs skip empty partitions