@BryanCutler 2019-06-21
This probably won't have an impact if data is not small, but given how many tests there are for these it should hopefully reduce some testing time.

Strangely, AggregateInPandasExec did not show an improvement, although the code is very similar. I'll have to look at that one in more detail.