@BryanCutler 2019-06-21
>do we have a test that process empty partitions?

We have one for pandas scalar udfs, but not specifically for these. Although I believe a bunch of empty partitions are created during the shuffle, it still would be good to test starting with one. Let me go ahead and add that.