@BryanCutler 2019-06-21
Looks like this cut almost 30s off of test time lol, happy friday @shaneknapp  !

GroupedMapPandasUDFTests
Ran 17 tests in 53.236s vs. 
Ran 18 tests in 24.102s