@HyukjinKwon 2019-06-22
Merged to master.