@shaneknapp 2019-06-24
haha, thanks!  ;)

On Fri, Jun 21, 2019 at 12:31 PM Bryan Cutler <notifications@github.com>
wrote:

> Looks like this cut almost 30s off of test time lol, happy friday
> @shaneknapp <https://github.com/shaneknapp> !
>
> GroupedMapPandasUDFTests
> Ran 17 tests in 53.236s vs.
> Ran 18 tests in 24.102s
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/apache/spark/pull/24926?email_source=notifications&email_token=AAMIHLG3EDINJ7FWBJQNPPLP3UUAVA5CNFSM4H2GLX62YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODYJMLKQ#issuecomment-504546730>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAMIHLE3ZXVDGLXV76ZHCO3P3UUAVANCNFSM4H2GLX6Q>
> .
>